requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
pymongo==4.6.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt>=4.0.1
